+++ src/main/kotlin/com/hellofresh/deblibs/BaseClient.kt
@@ -19,6 +19,8 @@ package com.hellofresh.deblibs
 import okhttp3.MediaType
 import okhttp3.OkHttpClient
 import okhttp3.Request
+import okhttp3.RequestBody
+import okhttp3.Response
 
 interface BaseClient : Runnable {
 
@@ -33,5 +35,28 @@ interface BaseClient : Runnable {
                     .addHeader("User-Agent", "deblibs-gradle-plugin")
                     .addHeader("Content-Type", "application/json")
             }
+    }
+
+    fun postRequest(
+        json: String,
+        url: String,
+        headerName: String,
+        headerValue: String
+    ): Pair<Response?, Int> {
+        val requestBody = RequestBody.create(JSON, json)
+        val request = requestBuilder.addHeader(headerName, headerValue)
+            .url(url)
+            .post(requestBody)
+            .build()
+        var response: Response? = null
+        val status: Int
+
+        try {
+            response = client.newCall(request).execute()
+            status = response.code()
+        } finally {
+            response?.close()
+        }
+        return Pair(response, status)
     }
 }